sched: freq: Reevaluate throttle if frequency requested changes
This is a bug that Andres found. Basically, if we're already in a down
throttle, any frequency up requests are ignored till the throttle period
ends. In this approach, we break out of any throttle if the requested
frequency changes and we check for the new request. Incase we are still
to throttle, we will continue once we return to the loop.

Test: uibench janktests
Bug: 33164692
Change-Id: Id5aa55da868e0322f96601c4f87199fb607a1efb
Signed-off-by: Joel Fernandes <joelaf@google.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>